Use a VERSION file for package version
Having to read __version__ from package can cause a race condition where
package can not be imported because of not yet installed dependencies
but needs to be imported to install those dependencies.

The 'file:' directive is not supported for version in setup.cfg using
setuptools (version 38.5.2). That's why the read_version_file.py is
required.

Update instructions has been added to the file in case setuptools adds
support for the directive.